// client/store/types/message.ts
export interface IMessage
{
    id: string;
    chat_id: string;
    user: IUser;
    text?: string;
    image?: string;
    audio?: string;
    video?: string;
    is_read?: boolean;
    is_edited?: boolean;
    is_forwarded?: boolean;
    reactions: IReaction[];
    created_at: Date;
    updated_at?: Date;
}

export interface IMessageState 
{ 
    message: IMessage | null;
    messages: IMessage[];
    error?: string;
}

export enum MessageActionTypes
{
    CREATE_MESSAGE = 'CREATE_MESSAGE',
    ASYNC_CREATE_MESSAGE = 'ASYNC_CREATE_MESSAGE',
    CREATE_MESSAGE_ERROR = 'CREATE_MESSAGE_ERROR',

    SET_MESSAGE = 'SET_MESSAGE',
    ASYNC_SET_MESSAGE = 'ASYNC_SET_MESSAGE',
    SET_MESSAGE_ERROR = 'SET_MESSAGE_ERROR',

    SET_ALL_MESSAGES = 'SET_ALL_MESSAGES',
    ASYNC_SET_ALL_MESSAGES = 'ASYNC_SET_ALL_MESSAGES',
    SET_ALL_MESSAGES_ERROR = 'SET_ALL_MESSAGES_ERROR',

    // UPDATE_MESSAGE = 'UPDATE_MESSAGE',
    // ASYNC_UPDATE_MESSAGE = 'ASYNC_UPDATE_MESSAGE',
    // UPDATE_MESSAGE_ERROR = 'UPDATE_MESSAGE_ERROR',

    SEARCH_MESSAGES = 'SEARCH_MESSAGES',
    ASYNC_SEARCH_MESSAGE = 'ASYNC_SEARCH_MESSAGE',
    SEARCH_MESSAGE_ERROR = 'SEARCH_MESSAGE_ERROR',

    DELETE_MESSAGE = 'DELETE_MESSAGE',
    ASYNC_DELETE_MESSAGE = 'ASYNC_DELETE_MESSAGE',
    DELETE_MESSAGE_ERROR = 'DELETE_MESSAGE_ERROR',

    DELETE_ALL_MESSAGES_IN_CHAT = 'DELETE_ALL_MESSAGES_IN_CHAT',
    ASYNC_DELETE_ALL_MESSAGES_IN_CHAT = 'ASYNC_DELETE_ALL_MESSAGES_IN_CHAT',
    DELETE_ALL_MESSAGES_IN_CHAT_ERROR = 'DELETE_ALL_MESSAGES_IN_CHAT_ERROR'
}